requirements.txt
streamlit
numpy
tensorflow
Pillow
